OCT-75: refactor Criteria type
