+++ components/catalogs/front/src/components/CatalogEdit/hooks/useCatalogCriteriaState.ts
@@ -1,16 +1,16 @@
-import {CriteriaState} from '../../ProductSelection/models/Criteria';
 import {useQuery} from 'react-query';
+import {AnyCriterionState} from '../../ProductSelection';
 
 type ResultError = Error | null;
 type Result = {
     isLoading: boolean;
     isError: boolean;
     error: ResultError;
-    data: CriteriaState | undefined;
+    data: AnyCriterionState[] | undefined;
 };
 
 export const useCatalogCriteriaState = (id: string): Result => {
-    return useQuery<CriteriaState, ResultError>(['catalog_criteria', id], async () => {
+    return useQuery<AnyCriterionState[], ResultError>(['catalog_criteria', id], async () => {
         const response = await fetch(`/rest/catalogs/${id}/data`, {
             headers: {
                 'X-Requested-With': 'XMLHttpRequest',
