+++ components/catalogs/front/src/components/ProductSelection/index.ts
@@ -1,2 +1,3 @@
 /* istanbul ignore file */
 export {ProductSelection} from './ProductSelection';
+export type {AnyCriterionState} from './models/Criteria';
