--- components/catalogs/front/src/components/ProductSelection/models/Criteria.d.ts
@@ -1,8 +1,0 @@
-import {StatusCriterionState} from '../criteria/StatusCriterion';
-import {Criterion} from './Criterion';
-
-export type AnyCriterionState = StatusCriterionState;
-
-export type CriteriaState = AnyCriterionState[];
-
-export type Criteria = Criterion<AnyCriterionState>[];
